+++ geotrek/common/parsers.py
@@ -588,3 +588,29 @@ class SitraParser(AttachmentParserMixin, Parser):
 
     def normalize_field_name(self, name):
         return name
+
+
+class OpenSystemParser(Parser):
+    url = 'http://proxy-xml.open-system.fr/rest.aspx'
+
+    def next_row(self):
+        params = {
+            'Login': self.login,
+            'Pass': self.password,
+            'Action': 'concentrateur_liaisons',
+        }
+        response = requests.get(self.url, params=params)
+        if response.status_code != 200:
+            raise GlobalImportError(_(u"Failed to download {url}. HTTP status code {status_code}").format(url=self.url, status_code=response.status_code))
+        self.root = ET.fromstring(response.content).find('Resultat').find('Objets')
+        self.nb = len(self.root)
+        for row in self.root:
+            id_sitra = row.find('ObjetCle').find('Cle').text
+            for liaison in row.find('Liaisons'):
+                yield {
+                    'id_sitra': id_sitra,
+                    'id_opensystem': liaison.find('ObjetOS').find('CodeUI').text,
+                }
+
+    def normalize_field_name(self, name):
+        return name