+++ geotrek/tourism/serializers.py
@@ -87,7 +87,7 @@ class TouristicContentSerializer(PicturesSerializerMixin, PublishableSerializerM
         fields = ('id', 'description', 'description_teaser', 'category',
                   'themes', 'contact', 'email', 'website', 'practical_info',
                   'type1', 'type2', 'touristic_contents', 'touristic_events',
-                  'treks', 'pois', 'source', 'approved') + \
+                  'treks', 'pois', 'source', 'approved', 'reservation_id') + \
             ZoningSerializerMixin.Meta.fields + \
             PublishableSerializerMixin.Meta.fields + \
             PicturesSerializerMixin.Meta.fields